removed applet dependencies
