+++ Portal/pom.xml
@@ -46,8 +46,6 @@ knowledge of the CeCILL-B license and that you accept its terms.
         <vip.application-importer.version>1.18</vip.application-importer.version>
         <vip.api.version>1.21</vip.api.version>
         <vip.application.version>1.21</vip.application.version>
-        <vip.datamanager-applet.version>1.16.2</vip.datamanager-applet.version>
-        <vip.gatelab-applet.version>1.16.2</vip.gatelab-applet.version>
         <vip.datamanager.version>1.21.1</vip.datamanager.version>
         <vip.gatelab.version>1.15</vip.gatelab.version>
         <vip.docs.version>1.15</vip.docs.version>
