+++ Portal/vip-gatelab/src/main/java/fr/insalyon/creatis/vip/gatelab/client/view/launch/LoadMacWindow.java
@@ -56,6 +56,6 @@ public class LoadMacWindow extends Window {
                 "MacUpload", 
                 GWT.getModuleBaseURL() + "/fileuploadservice", 
                 "gateParentFolderID", "macFileID", baseDir, "dataManagerUploadComplete",
-                true, false));
+                false, false));
     }
 }
